2022/day12/day12_test.go
package day12_test

import (
	"os"
	"testing"

	"day12"
)

func TestPartOneExampleInput(t *testing.T) {
	input, err := os.Open("./testdata/example.txt")
	if err != nil {
		t.Fatal(err)
	}

	want := 31
	got := day12.PartOne(input)

	if want != got {
		t.Errorf("wanted %d; got %d", want, got)
	}
}

func TestPartOneRealInput(t *testing.T) {
	input, err := os.Open("./testdata/input.txt")
	if err != nil {
		t.Fatal(err)
	}

	want := 423
	got := day12.PartOne(input)

	if want != got {
		t.Errorf("wanted %d; got %d", want, got)
	}
}

func TestPartTwoExampleInput(t *testing.T) {
	input, err := os.Open("./testdata/example.txt")
	if err != nil {
		t.Fatal(err)
	}

	want := 29
	got := day12.PartTwo(input)

	if want != got {
		t.Errorf("wanted %d; got %d", want, got)
	}
}

func TestPartTwoRealInput(t *testing.T) {
	input, err := os.Open("./testdata/input.txt")
	if err != nil {
		t.Fatal(err)
	}

	want := 416
	got := day12.PartTwo(input)

	if want != got {
		t.Errorf("wanted %d; got %d", want, got)
	}
}

func BenchmarkPartTwoRealInput(t *testing.B) {
	input, err := os.Open("./testdata/input.txt")
	if err != nil {
		t.Fatal(err)
	}

	t.ResetTimer()
	day12.PartTwo(input)
}
